# backend/rentigo-backend/src/main/java/cc/martincao/rentigo/rentigobackend/vehicle/model/Location.java
package cc.martincao.rentigo.rentigobackend.vehicle.model;

import jakarta.persistence.*;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.hibernate.annotations.CreationTimestamp;
import org.hibernate.annotations.UpdateTimestamp;

import java.math.BigDecimal;
import java.util.Date;

@Data
@NoArgsConstructor
@Entity
@Table(name = "location")
public class Location {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Integer id;

    @Column(nullable = false)
    private String city;

    @Column(name = "center_name", nullable = false)
    private String centerName;

    @Column(nullable = false)
    private String address;

    @Column(precision = 10, scale = 6)
    private BigDecimal lng;

    @Column(precision = 10, scale = 6)
    private BigDecimal lat;

    @CreationTimestamp
    @Column(name = "created_at", nullable = false, updatable = false)
    private Date createdAt;

    @UpdateTimestamp
    @Column(name = "updated_at", nullable = false)
    private Date updatedAt;
}